Add Description to function examples
